# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Erupe API
  description: REST API for the Erupe Monster Hunter Frontier server emulator.
  version: 2.0.0
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /v2/admin/users:
    get:
      summary: Search accounts
      operationId: adminSearchUsers
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: query, name: q, schema: { type: string }, description: Case-insensitive username substring }
        - { in: query, name: limit, schema: { type: integer, default: 50, maximum: 200 } }
      responses:
        "200":
          description: Matching accounts
          content:
            application/json:
              schema:
                type: object
                properties:
                  users: { type: array, items: { $ref: "#/components/schemas/AdminUser" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v2/admin/users/{id}:
    get:
      summary: Account detail with characters
      operationId: adminGetUser
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: integer } }]
      responses:
        "200":
          description: Account and its characters
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/AdminUser"
                  - type: object
                    properties:
                      characters: { type: array, items: { $ref: "#/components/schemas/Character" } }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      summary: Grant/revoke operator rights or set the rights bitfield
      operationId: adminUpdateUser
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: integer } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                op: { type: boolean }
                rights: { type: integer, description: Course bitfield, see wiki Enumerations }
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUser" }
        "400":
          description: Neither field given, or an operator removing their own op flag (`self_demotion`)
        "404": { $ref: "#/components/responses/NotFound" }
  /v2/admin/users/{id}/password:
    post:
      summary: Reset an account password
      operationId: adminSetPassword
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: integer } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string }
      responses:
        "204": { description: Password replaced }
        "404": { $ref: "#/components/responses/NotFound" }
  /v2/admin/users/{id}/ban:
    post:
      summary: Ban an account (permanent unless `expires` is given)
      operationId: adminBan
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: integer } }]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                expires: { type: string, format: date-time }
      responses:
        "204": { description: Banned }
        "400": { description: Expiry in the past, or banning yourself (`self_ban`) }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      summary: Lift a ban
      operationId: adminUnban
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: integer } }]
      responses:
        "204": { description: Unbanned (also when there was no ban) }
  /v2/admin/notices:
    get:
      summary: List login notices
      operationId: adminListNotices
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: query, name: all, schema: { type: string }, description: Any value includes expired notices }
      responses:
        "200":
          description: Notices
          content:
            application/json:
              schema:
                type: object
                properties:
                  notices: { type: array, items: { $ref: "#/components/schemas/Notice" } }
    post:
      summary: Add a login notice
      description: Shown at sign-in after the static `LoginNotices` from config.json, without a restart.
      operationId: adminCreateNotice
      tags: [admin]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string, maxLength: 4096, description: MHFML text }
                expires: { type: string, format: date-time }
      responses:
        "201":
          description: Created notice
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Notice" }
  /v2/admin/notices/{id}:
    delete:
      summary: Remove a login notice
      operationId: adminDeleteNotice
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: integer } }]
      responses:
        "204": { description: Removed }
        "404": { $ref: "#/components/responses/NotFound" }
  /v2/admin/distributions:
    get:
      summary: List gift-box distributions
      operationId: adminListDistributions
      tags: [admin]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Distributions (headers only)
          content:
            application/json:
              schema:
                type: object
                properties:
                  distributions: { type: array, items: { $ref: "#/components/schemas/Distribution" } }
    post:
      summary: Create a distribution with its items
      description: |
        `characterId` null = every character. `type` and `items[].itemType`
        follow the Distribution enumerations on the wiki. `eventName`,
        `description` and `timesAcceptable` default to the game's GM-gift values.
      operationId: adminCreateDistribution
      tags: [admin]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/Distribution"
                - type: object
                  required: [items]
                  properties:
                    items: { type: array, minItems: 1, items: { $ref: "#/components/schemas/DistributionItem" } }
      responses:
        "201":
          description: Created distribution with its items
        "400": { description: No items, or a non-positive quantity }
  /v2/admin/distributions/{id}:
    get:
      summary: Distribution with its items
      operationId: adminGetDistribution
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: integer } }]
      responses:
        "200": { description: Distribution and items }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      summary: Delete a distribution and its items
      operationId: adminDeleteDistribution
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: integer } }]
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }
  /v2/admin/events:
    get:
      summary: List event rows (festa, diva, vs, mezfes)
      operationId: adminListEvents
      tags: [admin]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: object
                properties:
                  events: { type: array, items: { $ref: "#/components/schemas/AdminEvent" } }
  /v2/admin/events/{type}:
    put:
      summary: Start (or restart) an event cycle
      description: |
        Replaces every row of that type with one starting at `startTime`
        (default now). The game derives the whole schedule from that row, so
        this is how a Diva Defense or Festa cycle is restarted.
      operationId: adminStartEvent
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters: [{ in: path, name: type, required: true, schema: { type: string, enum: [festa, diva, vs, mezfes] } }]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                startTime: { type: string, format: date-time }
      responses:
        "200":
          description: The new event row
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminEvent" }
        "400": { description: Unknown event type }
    delete:
      summary: Stop an event (remove its rows)
      operationId: adminStopEvent
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters: [{ in: path, name: type, required: true, schema: { type: string, enum: [festa, diva, vs, mezfes] } }]
      responses:
        "200":
          description: Number of rows removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  removed: { type: integer }
  /v2/admin/content/reload:
    post:
      summary: Synchronise the content directory into the database
      description: |
        Re-reads `<ContentPath>/<table>/*.json` (default `game-data/content/`)
        and makes each table match its files: rows are matched on the file's
        `key`, updated in place, inserted, and — inside the file's `scope` —
        deleted when no longer listed. Each file is one transaction; on an
        error the files applied before it stay applied and are listed.
        Shop rows are read per request, so the change is live at once.
      operationId: adminReloadContent
      tags: [admin]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Every file was synchronised
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ContentReload" }
        "404": { description: The content directory does not exist }
        "422":
          description: A file could not be applied; `message` names it, `files` lists what was applied before
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ContentReload" }
  /v2/admin/config:
    get:
      summary: Read the config file as the editor shows it
      description: |
        Every option of the loaded config file (config.json) with its
        description, value on disk, default, and whether a saved change
        still waits for a restart. Secret options (database password, bot
        token, CapLink key) carry no `value` or `default`. `unknownKeys`
        lists keys Erupe ignores, such as misspelt options. The operator page
        at `/admin` is built on this endpoint.
      operationId: adminGetConfig
      tags: [admin]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: The config file
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConfigState" }
        "404": { description: The server was not started from a config file }
    patch:
      summary: Change options in the config file
      description: |
        Validates each value against its option (type, range, allowed
        values), checks that the resulting file still loads, keeps the
        previous file as `<file>.bak` and writes the new one. Keys the request
        does not name keep their spelling, order and value. Changes apply at
        the next restart. Every change is logged with the operator's ID
        (secret values are not logged).
      operationId: adminPatchConfig
      tags: [admin]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                version:
                  type: string
                  description: The `version` from GET; the change is refused if the file changed since
                set:
                  type: object
                  additionalProperties: true
                  description: Option path (e.g. `GameplayOptions.MaximumRP`) to its new JSON value
                  example: { "GameplayOptions.MaximumRP": 60000, "HideLoginNotice": false }
                unset:
                  type: array
                  items: { type: string }
                  description: Options to remove from the file, so they fall back to their default
      responses:
        "200":
          description: Written; the new state of the file
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConfigState" }
        "400": { description: Unknown option, invalid value, or nothing to change }
        "404": { description: The server was not started from a config file }
        "409": { description: The file changed since `version` (`config_changed`), or it is not JSON (`not_json`) }
        "422": { description: The resulting file would not load }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Session token returned by /v2/login or /v2/register
...
  responses:
    Unauthorized:
      description: Missing or invalid Bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: unauthorized
            message: Invalid or expired token
...
    Forbidden:
      description: Operator rights required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: No such resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    AdminUser:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        rights: { type: integer }
        op: { type: boolean }
        lastLogin: { type: string, format: date-time, nullable: true }
        discordId: { type: string, nullable: true }
        psnId: { type: string, nullable: true }
        banned: { type: boolean }
        banExpires: { type: string, format: date-time, nullable: true, description: null for a permanent ban or no ban }
    Notice:
      type: object
      properties:
        id: { type: integer }
        body: { type: string }
        createdAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time, nullable: true }
    Distribution:
      type: object
      properties:
        id: { type: integer, readOnly: true }
        characterId: { type: integer, nullable: true }
        type: { type: integer }
        deadline: { type: string, format: date-time, nullable: true }
        eventName: { type: string }
        description: { type: string }
        timesAcceptable: { type: integer }
        minHr: { type: integer, nullable: true }
        maxHr: { type: integer, nullable: true }
        minSr: { type: integer, nullable: true }
        maxSr: { type: integer, nullable: true }
        minGr: { type: integer, nullable: true }
        maxGr: { type: integer, nullable: true }
        rights: { type: integer, nullable: true }
        selection: { type: boolean, nullable: true }
...
    Character:
      type: object
      required: [id, name, isFemale, weapon, hr, gr, lastLogin, returning]
      properties:
        id:
          type: integer
          format: uint32
        name:
          type: string
        isFemale:
          type: boolean
        weapon:
          type: integer
          format: uint32
        hr:
          type: integer
          format: uint32
        gr:
          type: integer
          format: uint32
        lastLogin:
          type: integer
          format: int32
          description: Unix timestamp of last login
        returning:
          type: boolean
          description: True if character has not logged in for 90+ days
...